excel_parse/requirements.txt
DSPy
pandas
pyarrow
polars
duckdb
scikit-learn
sentence-transformers
python-dotenv
openpyxl
pytest
